Angular_Flask_API/flask_api/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
Werkzeug==3.0.2
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.3.0
